+++ packages/rum-core/test/common/apm-server.spec.js
@@ -33,6 +33,7 @@ import { captureBreakdown } from '../../src/performance-monitoring/breakdown'
 import { createServiceFactory } from '../'
 
 const { agentConfig, testConfig } = getGlobalConfig('rum-core')
+import { LOCAL_CONFIG_KEY } from '../../src/common/constants'
 
 function generateTransaction(count, breakdown = false) {
   var result = []
@@ -408,38 +409,64 @@ describe('ApmServer', function() {
 
   it('should ndjson metricsets along with transactions', function() {
     const tr = generateTransaction(1, true)
-    jasmine.clock().install()
-    jasmine.clock().mockDate(new Date(0))
     const payload = performanceMonitoring.convertTransactionsToServerModel(tr)
     const result = apmServer.ndjsonTransactions(payload)
     /* eslint-disable max-len */
     const expected = [
       '{"transaction":{"id":"transaction-id-0","trace_id":"trace-id-0","name":"transaction #0","type":"transaction","duration":990,"context":{"page":{"referer":"referer","url":"url"}},"span_count":{"started":1},"sampled":true}}\n',
       '{"span":{"id":"span-id-0-1","transaction_id":"transaction-id-0","parent_id":"transaction-id-0","trace_id":"trace-id-0","name":"name","type":"type","sync":false,"start":10,"duration":10}}\n',
-      '{"metricset":{"timestamp":0,"transaction":{"name":"transaction #0","type":"transaction"},"samples":{"transaction.duration.count":{"value":1},"transaction.duration.sum.us":{"value":990},"transaction.breakdown.count":{"value":1}}}}\n',
-      '{"metricset":{"timestamp":0,"transaction":{"name":"transaction #0","type":"transaction"},"span":{"type":"app"},"samples":{"span.self_time.count":{"value":1},"span.self_time.sum.us":{"value":980}}}}\n',
-      '{"metricset":{"timestamp":0,"transaction":{"name":"transaction #0","type":"transaction"},"span":{"type":"type"},"samples":{"span.self_time.count":{"value":1},"span.self_time.sum.us":{"value":10}}}}\n'
+      '{"metricset":{"transaction":{"name":"transaction #0","type":"transaction"},"samples":{"transaction.duration.count":{"value":1},"transaction.duration.sum.us":{"value":990},"transaction.breakdown.count":{"value":1}}}}\n',
+      '{"metricset":{"transaction":{"name":"transaction #0","type":"transaction"},"span":{"type":"app"},"samples":{"span.self_time.count":{"value":1},"span.self_time.sum.us":{"value":980}}}}\n',
+      '{"metricset":{"transaction":{"name":"transaction #0","type":"transaction"},"span":{"type":"type"},"samples":{"span.self_time.count":{"value":1},"span.self_time.sum.us":{"value":10}}}}\n'
     ].join('')
     expect(result).toEqual([expected])
-    jasmine.clock().uninstall()
   })
 
   if (isVersionInRange(testConfig.stackVersion, '7.3.0')) {
     it('should fetch remote config', async () => {
+      spyOn(configService, 'setLocalConfig')
+      spyOn(configService, 'getLocalConfig')
+
       var config = await apmServer.fetchConfig('nonexistent-service')
-      expect(config).toEqual({})
+      expect(configService.getLocalConfig).toHaveBeenCalled()
+      expect(configService.setLocalConfig).toHaveBeenCalled()
+
+      expect(config).toEqual({ etag: jasmine.any(String) })
 
       config = await apmServer.fetchConfig(
         'nonexistent-service',
         'nonexistent-env'
       )
-      expect(config).toEqual({})
+      expect(config).toEqual({ etag: jasmine.any(String) })
 
       try {
         config = await apmServer.fetchConfig()
       } catch (e) {
         expect(e).toBe('serviceName is required for fetching central config.')
       }
+    })
+
+    it('should use local config if available', async () => {
+      configService.setLocalConfig({
+        transaction_sample_rate: '0.5',
+        etag: 'test'
+      })
+
+      apmServer._makeHttpRequest = () => {
+        return Promise.resolve({
+          status: 304
+        })
+      }
+
+      let config = await apmServer.fetchConfig(
+        'nonexistent-service',
+        'nonexistent-env'
+      )
+      expect(config).toEqual({
+        transaction_sample_rate: '0.5',
+        etag: 'test'
+      })
+      sessionStorage.removeItem(LOCAL_CONFIG_KEY)
     })
   }
 })