+++ packages/rum-core/test/performance-monitoring/performance-monitoring.spec.js
@@ -203,39 +203,39 @@ describe('PerformanceMonitoring', function() {
       transactionDurationThreshold: 200
     })
     spyOn(logger, 'debug').and.callThrough()
-    const transaction1 = new Transaction()
+    const transaction1 = new Transaction('test', 'custom', { id: 1 })
     transaction1.end()
     transaction1._start = 0
     transaction1._end = 201
     expect(transaction1.duration()).toBe(201)
     expect(performanceMonitoring.filterTransaction(transaction1)).toBe(false)
     expect(logger.debug).toHaveBeenCalledWith(
-      'Transaction was discarded! Transaction duration (201) is greater than the transactionDurationThreshold configuration (200)'
+      'transaction(1, test) was discarded! Transaction duration (201) is greater than the transactionDurationThreshold configuration (200)'
     )
     logger.debug.calls.reset()
 
-    const transaction2 = new Transaction()
+    const transaction2 = new Transaction('test2', 'custom', { id: 2 })
     transaction2.end()
     transaction2._end = transaction2._end + 100
     expect(performanceMonitoring.filterTransaction(transaction2)).toBe(false)
     expect(logger.debug).toHaveBeenCalledWith(
-      'Transaction was discarded! Transaction does not include any spans'
+      'transaction(2, test2) was discarded! Transaction does not include any spans'
     )
     logger.debug.calls.reset()
 
-    const transaction3 = new Transaction()
+    const transaction3 = new Transaction(null, null, { id: 3 })
     expect(performanceMonitoring.filterTransaction(transaction3)).toBe(false)
     expect(logger.debug).toHaveBeenCalledWith(
-      "Transaction was discarded! Transaction wasn't ended"
+      "transaction(3, Unknown) was discarded! Transaction wasn't ended"
     )
     logger.debug.calls.reset()
 
-    const transaction4 = new Transaction()
+    const transaction4 = new Transaction('', '', { id: 4 })
     transaction4.end()
     transaction4._start = transaction4._end = 0
     expect(performanceMonitoring.filterTransaction(transaction4)).toBe(false)
     expect(logger.debug).toHaveBeenCalledWith(
-      'Transaction was discarded! Transaction duration is 0'
+      'transaction(4, Unknown) was discarded! Transaction duration is 0'
     )
   })
 
@@ -248,6 +248,7 @@ describe('PerformanceMonitoring', function() {
     spyOn(logger, 'debug').and.callThrough()
     expect(logger.debug).not.toHaveBeenCalled()
     var tr = new Transaction('transaction', 'transaction', {
+      id: 212,
       transactionSampleRate: 1,
       managed: true,
       checkBrowserResponsiveness: true
@@ -262,7 +263,7 @@ describe('PerformanceMonitoring', function() {
     var wasBrowserResponsive = performanceMonitoring.filterTransaction(tr)
     expect(wasBrowserResponsive).toBe(false)
     expect(logger.debug).toHaveBeenCalledWith(
-      'Transaction was discarded! Browser was not responsive enough during the transaction.',
+      'transaction(212, transaction) was discarded! Browser was not responsive enough during the transaction.',
       ' duration:',
       3000,
       ' browserResponsivenessCounter:',
@@ -296,62 +297,6 @@ describe('PerformanceMonitoring', function() {
     }, 10)
   })
 
-  it('should adjust transaction start based on earliest span start', function(done) {
-    const firstSpan = new Span('first-span-name', 'first-span')
-    firstSpan.end()
-
-    const transaction = new Transaction('/', 'transaction', {})
-    transaction.onEnd = function() {
-      performanceMonitoring.adjustTransactionTime(transaction)
-      expect(transaction._start).toBe(firstSpan._start)
-      expect(transaction._end).toBeGreaterThanOrEqual(lastSpan._end)
-      done()
-    }
-    transaction.spans.push(firstSpan)
-
-    const lastSpan = transaction.startSpan('last-span-name', 'last-span')
-    lastSpan.end()
-
-    transaction.detectFinish()
-  })
-
-  it('should adjust transaction end based on latest span end', function(done) {
-    const transaction = new Transaction('/', 'transaction', {})
-    const transactionStart = transaction._start
-
-    const firstSpan = transaction.startSpan('first-span-name', 'first-span')
-    firstSpan.end()
-
-    const longSpan = transaction.startSpan('long-span-name', 'long-span')
-
-    const lastSpan = transaction.startSpan('last-span-name', 'last-span')
-    lastSpan.end()
-
-    longSpan.end()
-    longSpan.end += 500
-
-    transaction.onEnd = function() {
-      performanceMonitoring.adjustTransactionTime(transaction)
-      expect(transaction._start).toBe(transactionStart)
-      expect(transaction._end).toBeGreaterThanOrEqual(longSpan._end)
-      done()
-    }
-    transaction.detectFinish()
-  })
-
-  it('should truncate active spans after transaction ends', () => {
-    const transaction = new Transaction('transaction', 'transaction')
-    const span = transaction.startSpan('test', 'test')
-    expect(transaction.spans.length).toBe(0)
-    expect(Object.keys(transaction._activeSpans).length).toBe(1)
-    transaction.end()
-
-    performanceMonitoring.adjustTransactionTime(transaction)
-    expect(transaction.spans.length).toBe(1)
-    expect(Object.keys(transaction._activeSpans).length).toBe(0)
-    expect(span.type).toContain('.truncated')
-  })
-
   it('should create correct payload', function() {
     var tr = new Transaction('transaction1', 'transaction1type', {
       transactionSampleRate: 1
@@ -362,9 +307,7 @@ describe('PerformanceMonitoring', function() {
     tr.detectFinish()
 
     expect(tr._end).toBeDefined()
-    if (tr._end === tr._start) {
-      tr._end = tr._end + 100
-    }
+    tr._end = span._end + 100
 
     var payload = performanceMonitoring.createTransactionPayload(tr)
     expect(payload.name).toBe('transaction1')
@@ -374,21 +317,6 @@ describe('PerformanceMonitoring', function() {
     expect(payload.spans[0].type).toBe('span1type')
     expect(payload.spans[0].start).toBe(span._start - tr._start)
     expect(payload.spans[0].duration).toBe(span._end - span._start)
-  })
-
-  it('should not produce negative durations while adjusting to the spans', function() {
-    var transaction = new Transaction('transaction', 'transaction')
-    var span = transaction.startSpan('test', 'test')
-    span.end()
-    span._end += 100
-    span = transaction.startSpan('test', 'external.http')
-
-    span.end()
-    span._start = 10000000
-    span._end = 11000000
-    transaction.end()
-    performanceMonitoring.adjustTransactionTime(transaction)
-    expect(span.duration()).toBe(0)
   })
 
   it('should sendPageLoadMetrics', function(done) {
@@ -670,11 +598,11 @@ describe('PerformanceMonitoring', function() {
             },
             {
               event: 'invoke',
-              source: XMLHTTPREQUEST
+              source: FETCH
             },
             {
               event: 'invoke',
-              source: FETCH
+              source: XMLHTTPREQUEST
             }
           ])
           done()