+++ packages/rum-vue/test/e2e/components/Lazy.vue
@@ -1,0 +1,5 @@
+<template>
+  <div>
+    <h2>Lazy loaded component</h2>
+  </div>
+</template>