+++ packages/rum/src/apm-base.js
@@ -43,7 +43,7 @@ class ApmBase {
       /**
        * Set Agent version to be sent as part of metadata to the APM Server
        */
-      configService.setVersion('4.5.1')
+      configService.setVersion('4.6.0')
       this.config(config)
       const loggingService = this.serviceFactory.getService('LoggingService')
       /**
@@ -111,14 +111,14 @@ class ApmBase {
           } else {
             const { key, value, allowed } = invalid[0]
             loggingService.warn(
-              `Invalid value "${value}" for ${key}. Allowed: ${allowed}.`
+              `invalid value "${value}" for ${key}. Allowed: ${allowed}.`
             )
           }
         }
         return config
       })
       .catch(error => {
-        loggingService.warn('Failed fetching config:', error)
+        loggingService.warn('failed fetching config:', error)
       })
   }
 
@@ -186,10 +186,10 @@ class ApmBase {
     } else {
       const loggingService = this.serviceFactory.getService('LoggingService')
       const separator = ', '
-      let message = "RUM Agent isn't correctly configured: "
+      let message = "RUM agent isn't correctly configured. "
 
       if (missing.length > 0) {
-        message += 'Missing config - ' + missing.join(separator)
+        message += missing.join(separator) + ' is missing'
         if (invalid.length > 0) {
           message += separator
         }